# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniSync",
    "version": "see VERSION"
  },
  "paths": {
...
    "/config": {
      "get": {
        "summary": "Get Config",
        "description": "Return the current global configuration.",
        "operationId": "get_config_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalConfigResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Config",
        "description": "Update global configuration with partial values.",
        "operationId": "update_config_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalConfigResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/config/test-sync": {
      "post": {
        "summary": "Test Sync",
        "description": "Run a round-trip sync test with a dummy file.",
        "operationId": "test_sync_config_test_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestSyncResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, for people. Show this text."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable code (snake_case) to branch on."
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Extra data for some codes, e.g. errors, invalid_paths, names, replacement, retry_after."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "The request's id (also the X-Request-ID header); the server log lines of the request carry it."
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "The body of every error answer (4xx and 5xx); docs/api-errors.md lists the codes."
      },
...
      "GlobalConfigResponse": {
        "properties": {
          "log_level": {
            "type": "string",
            "title": "Log Level",
            "default": "INFO"
          },
          "history_days": {
            "type": "integer",
            "title": "History Days",
            "default": 90
          }
        },
        "type": "object",
        "title": "GlobalConfigResponse"
      },
      "GlobalConfigUpdateRequest": {
        "properties": {
          "log_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "DEBUG",
                  "INFO",
                  "WARNING",
                  "ERROR",
                  "CRITICAL"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Level"
          },
          "history_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3650.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "History Days"
          }
        },
        "type": "object",
        "title": "GlobalConfigUpdateRequest"
      },
...
      "TestSyncRequest": {
        "properties": {
          "local_dir": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Local Dir"
          },
          "remote_dir": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Remote Dir"
          }
        },
        "type": "object",
        "required": [
          "local_dir",
          "remote_dir"
        ],
        "title": "TestSyncRequest",
        "description": "POST /config/test-sync: the same checks as a new profile's folders."
      },
      "TestSyncResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps",
            "default": []
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "TestSyncResponse"
      },
...
    }
  }
}